requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
ollama>=0.3.0
requests>=2.31.0
